requirements.txt
pyOpenSSL>=23.1.1
construct>=2.9.29
future
pyasn1
six
biplist
hexdump
